extract 2.1 int tests
